Converted all size2 stuff to intsize2.
git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@1414 146ae76c-2b19-ef4c-9a69-67eb2b888ad7

+++ MyGUIPlugin/Plugin/Widgets/Timeline/TimelineMarker.cs
@@ -118,9 +118,9 @@ namespace MyGUIPlugin
             Time = newTime;
         }
 
-        void timelineView_CanvasHeightChanged(float newSize)
+        void timelineView_CanvasHeightChanged(int newSize)
         {
-            widget.setSize(widget.Width, (int)newSize);
+            widget.setSize(widget.Width, newSize);
         }
 
         void timelineView_PixelsPerSecondChanged(object sender, EventArgs e)
